tested exported checkpoints on 0211

+++ modelopt/torch/export/unified_export_hf.py
@@ -525,7 +525,7 @@ def _export_quantized_weight(
             )[0]
 
         quantized_weight = to_quantized_weight(
-            weight.to(dtype),
+            weight.to(torch.bfloat16),
             weight_scale,
             quantization_format,
             weight_scale_2,
@@ -542,7 +542,7 @@ def _export_quantized_weight(
         )
 
         quantized_weight = to_quantized_weight(
-            weight.to(dtype),
+            weight.to(torch.bfloat16),
             weight_scale,
             quantization_format,
             weight_scale_2,
